Implementing user authentication

+++ src/main/java/br/dev/mhc/financialassistantapi/services/UserService.java
@@ -9,21 +9,39 @@ import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort.Direction;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import br.dev.mhc.financialassistantapi.dto.UserDTO;
+import br.dev.mhc.financialassistantapi.dto.UserNewDTO;
 import br.dev.mhc.financialassistantapi.entities.User;
+import br.dev.mhc.financialassistantapi.entities.enums.Profile;
 import br.dev.mhc.financialassistantapi.repositories.UserRepository;
+import br.dev.mhc.financialassistantapi.security.UserSpringSecurity;
 import br.dev.mhc.financialassistantapi.services.email.EmailService;
+import br.dev.mhc.financialassistantapi.services.exceptions.AuthorizationException;
 import br.dev.mhc.financialassistantapi.services.exceptions.DataIntegrityException;
 import br.dev.mhc.financialassistantapi.services.exceptions.ObjectNotFoundException;
 
 @Service
 public class UserService {
 
+	public static UserSpringSecurity authenticated() {
+		try {
+		return (UserSpringSecurity) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+		} catch (Exception e) {
+			return null;
+		}
+	}
+
+	@Autowired
+	private BCryptPasswordEncoder pe;
+
 	@Autowired
 	private UserRepository repository;
-	
+
 	@Autowired
 	private EmailService emailService;
 
@@ -40,6 +58,11 @@ public class UserService {
 
 	@Transactional
 	public User update(User obj) {
+		UserSpringSecurity userSS = UserService.authenticated();
+		if(userSS==null || !userSS.hasRole(Profile.ADMIN) && !obj.getId().equals(userSS.getId())) {
+			throw new AuthorizationException("Access denied");
+		}
+
 		User newObj = findById(obj.getId());
 		updateData(newObj, obj);
 		return repository.save(newObj);
@@ -51,6 +74,11 @@ public class UserService {
 	}
 
 	public void delete(Long id) {
+		UserSpringSecurity userSS = UserService.authenticated();
+		if(userSS==null || !userSS.hasRole(Profile.ADMIN) && !id.equals(userSS.getId())) {
+			throw new AuthorizationException("Access denied");
+		}
+
 		findById(id);
 		try {
 			repository.deleteById(id);
@@ -69,6 +97,11 @@ public class UserService {
 	}
 
 	public User findById(Long id) {
+		UserSpringSecurity userSS = UserService.authenticated();
+		if(userSS==null || !userSS.hasRole(Profile.ADMIN) && !id.equals(userSS.getId())) {
+			throw new AuthorizationException("Access denied");
+		}
+		
 		Optional<User> obj = repository.findById(id);
 		return obj.orElseThrow(
 				() -> new ObjectNotFoundException("Object not found! Id: " + id + ", Type: " + User.class.getName()));
@@ -76,5 +109,16 @@ public class UserService {
 
 	public User findByEmail(String email) {
 		return repository.findByEmail(email);
+	}
+
+	public User fromDTO(UserDTO objDto) {
+		return new User(objDto.getId(), objDto.getNickname(), objDto.getEmail(), null, objDto.getRegistrationMoment(),
+				objDto.getLastAccess(), objDto.isActive());
+	}
+
+	public User fromDTO(UserNewDTO objDTO) {
+		User user = new User(null, objDTO.getNickname(), objDTO.getEmail(), pe.encode(objDTO.getPassword()), null, null,
+				true);
+		return user;
 	}
 }
